separate out utils for clearer testing
create alias for src folder so we can use absolute paths
move toMachine helper method to utils
move friendlyEventName method to utils
change StateChartNode from a class component to a functional component
