+++ src/lib/utils/flatten/index.ts
@@ -1,0 +1,3 @@
+export function flatten<T>(array: T[][]): T[] {
+  return ([] as T[]).concat(...array)
+}
